move some more stuff to Harmony
