+++ scopes/component/merging/index.ts
@@ -1,6 +1,6 @@
 import { MergingAspect } from './merging.aspect';
 
-export { getComponentStatus, merge, ComponentStatus } from './merge-snaps';
+export { mergeReport } from './merge-cmd';
 export type { MergingMain } from './merging.main.runtime';
 export default MergingAspect;
 export { MergingAspect };
