+++ src/consumer/versions-ops/merge-version/index.ts
@@ -1,3 +1,9 @@
 export { default as threeWayMerge } from './three-way-merge';
 export { mergeVersion, FileStatus, getMergeStrategyInteractive, getMergeStrategy, MergeOptions } from './merge-version';
-export { MergeStrategy, ApplyVersionResults, ApplyVersionResult, FailedComponents } from './merge-version';
+export {
+  MergeStrategy,
+  ApplyVersionResults,
+  ApplyVersionResult,
+  FailedComponents,
+  applyVersionReport,
+} from './merge-version';
